refactor: 컴퓨터의 랜덤 숫자 고르는 로직 분리
